Changes for assignment screen
