+++ src/Kanbas/Courses/Assignments/index.js
@@ -7,9 +7,10 @@ import {
 import "./index.css"
 import {Link, useNavigate, useParams} from "react-router-dom";
 import {RxDragHandleDots2} from "react-icons/rx";
-import React from "react";
+import {useEffect} from "react";
 import {useDispatch, useSelector} from "react-redux";
-import {deleteAssignment} from "./assignmentsReducer";
+import {deleteAssignment, setAssignments} from "./assignmentsReducer";
+import {deleteAssignmentFromDB, getAssignments} from "./client";
 
 function Assignments() {
   const {courseId} = useParams();
@@ -26,9 +27,19 @@ function Assignments() {
 
   const deleteAss = (assignmentId) => {
     if (window.confirm('Are you sure you want to delete this assignment?')) {
+      deleteAssignmentFromDB(courseId, assignmentId).then(() => {
+        dispatch(
+            deleteAssignment({courseId: courseId, assignmentId: assignmentId}));
+      });
       dispatch(deleteAssignment(assignmentId));
     }
   }
+
+  useEffect(() => {
+    getAssignments(courseId).then((assignments) => {
+      dispatch(setAssignments(assignments));
+    });
+  }, [courseId]);
 
   return (
       <>
@@ -80,7 +91,7 @@ function Assignments() {
                     <div className="col-9 align-self-center">
                       <Link
                           key={assignment._id}
-                          to={`/Kanbas/Courses/${courseId}/Assignments/${assignment._id}`}>
+                          to={`/Kanbas/Courses/${courseId}/EditAssignment/${assignment._id}`}>
                         {assignment.title}
                       </Link>
                       {assignment.description && <p
