+++ src/Kanbas/Courses/index.js
@@ -5,9 +5,8 @@ import Home from "./Home";
 import Modules from "./Modules";
 import Assignments from "./Assignments";
 import Grades from "./Grades";
-import AssignmentEditor from "./Assignments/AssignmentEditor";
 import CourseNavigation from "./CourseNavigation";
-import AddAssignment from "./Assignments/AddAssignment";
+import AddEditAssignment from "./Assignments/AddEditAssignment";
 import {useEffect, useState} from "react";
 import axios from "axios";
 
@@ -22,7 +21,7 @@ function Courses() {
   };
   useEffect(() => {
     findCourseById(courseId);
-  }, []);
+  }, [courseId]);
 
   return (
       <>
@@ -42,12 +41,9 @@ function Courses() {
               <Route path="Home" element={<Home/>}/>
               <Route path="Modules" element={<Modules/>}/>
               <Route path="Assignments" element={<Assignments/>}/>
-              <Route
-                  path="Assignments/:assignmentId"
-                  element={<AssignmentEditor/>}
-              />
               <Route path="Grades" element={<Grades/>}/>
-              <Route path="AddAssignment" element={<AddAssignment/>}/>
+              <Route path="AddAssignment" element={<AddEditAssignment/>}/>
+              <Route path="EditAssignment/:assignmentId" element={<AddEditAssignment/>}/>
             </Routes>
           </div>
         </div>
